requirements.txt
numpy
scipy
pulser-core
pulser-simulation
myqlm>=1.7.0